+++ projects/js-packages/charts/src/components/conversion-funnel-chart/index.ts
@@ -6,4 +6,4 @@ export type {
 	StepRateRenderProps,
 	MainMetricRenderProps,
 	TooltipRenderProps,
-} from './conversion-funnel-chart';
+} from './types';
